+++ .vscode/launch.json
@@ -2,26 +2,10 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Debug Bayes",
-            "type": "debugpy",
-            "request": "launch",
-            "module": "bayesian_optimization",
-            "console": "integratedTerminal"
-        },
-        {
-            "name": "Run Bayes",
-            "type": "debugpy",
-            "request": "launch",
-            "module": "bayesian_optimization",
-            "console": "integratedTerminal",
-            "justMyCode": false,
-            "stopOnEntry": false
-        },
-        {
             "name": "Debug BayesMOO",
             "type": "debugpy",
             "request": "launch",
-            "module": "bayesian_moo",
+            "module": "bayesian_optimization",
             "console": "integratedTerminal",
             "env": {
                 "BAYESIAN_DEBUG": "true"
@@ -31,7 +15,7 @@
             "name": "Run BayesMOO",
             "type": "debugpy",
             "request": "launch",
-            "module": "bayesian_moo",
+            "module": "bayesian_optimization",
             "console": "integratedTerminal",
             "justMyCode": false,
             "stopOnEntry": false
